Rename of package, class, and method names
> gracenote.object to gracenote.bean
> GraceDaoImpl to DatabaseGraceDao (with test class)
> OpeningDaoImpl to DatabaseOpeningDao (with test class)
> README proper use of PUT rather than POST
