+++ src/main/java/io/github/jmgloria07/gracenote/dao/OpeningDao.java
@@ -2,7 +2,7 @@ package io.github.jmgloria07.gracenote.dao;
 
 import java.util.List;
 
-import io.github.jmgloria07.gracenote.object.Opening;
+import io.github.jmgloria07.gracenote.bean.Opening;
 
 public interface OpeningDao {
 	public List<Opening> getOpenings();
