+++ src/main/java/io/github/jmgloria07/gracenote/dao/repository/OpeningRepository.java
@@ -2,7 +2,7 @@ package io.github.jmgloria07.gracenote.dao.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import io.github.jmgloria07.gracenote.object.Opening;
+import io.github.jmgloria07.gracenote.bean.Opening;
 
 public interface OpeningRepository extends JpaRepository<Opening, Long> {
