+++ src/main/java/io/github/jmgloria07/gracenote/dao/repository/UserRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import io.github.jmgloria07.gracenote.object.User;
+import io.github.jmgloria07.gracenote.bean.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
