+++ src/main/java/io/github/jmgloria07/gracenote/service/OpeningService.java
@@ -2,7 +2,7 @@ package io.github.jmgloria07.gracenote.service;
 
 import java.util.List;
 
-import io.github.jmgloria07.gracenote.object.Opening;
+import io.github.jmgloria07.gracenote.bean.Opening;
 
 public interface OpeningService {
 	public List<Opening> getOpenings();
