+++ src/main/java/io/github/jmgloria07/gracenote/service/OpeningServiceImpl.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import io.github.jmgloria07.gracenote.bean.Opening;
 import io.github.jmgloria07.gracenote.dao.OpeningDao;
-import io.github.jmgloria07.gracenote.object.Opening;
 
 @Service
 public class OpeningServiceImpl implements OpeningService {
