+++ src/test/java/io/github/jmgloria07/gracenote/dao/DatabaseOpeningDaoTest.java
@@ -10,19 +10,19 @@ import static org.mockito.Mockito.*;
 import java.util.ArrayList;
 import java.util.Optional;
 
+import io.github.jmgloria07.gracenote.bean.Opening;
+import io.github.jmgloria07.gracenote.bean.exception.GraceNoteNotFoundException;
 import io.github.jmgloria07.gracenote.dao.repository.OpeningRepository;
-import io.github.jmgloria07.gracenote.object.Opening;
-import io.github.jmgloria07.gracenote.object.exception.GraceNoteNotFoundException;
 
-public class OpeningDaoImplTest {
-	private static OpeningDaoImpl unit;
+public class DatabaseOpeningDaoTest {
+	private static DatabaseOpeningDao unit;
 	
 	private static final long MOCK_OPENING_ID = 1L;
 	private static final long MOCK_MISSING_OPENING_ID = 2L;
 	
 	@BeforeAll
 	public static void init() {
-		unit = new OpeningDaoImpl();
+		unit = new DatabaseOpeningDao();
 		
 		unit.openingRepository = mock(OpeningRepository.class);
 		doReturn(new ArrayList<Opening>()).when(unit.openingRepository).findAll();
